missing tables

+++ src/main/java/com/upn/restaurant/service/UsuarioService.java
@@ -2,6 +2,6 @@ package com.upn.restaurant.service;
 
 import com.upn.restaurant.model.entity.Usuario;
 
-public interface UsuarioService extends CrudService<Usuario, Integer>{
+public interface UsuarioService extends CrudService<Usuario, Long>{
 
 }
